move printing code to asm_ostream.cpp and unify
